requirements.txt
appdirs==1.4.4
asgiref==3.8.1
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
-e git+https://github.com/atiqullah-ahmadzai/sil1-task-4.git@948deb6c322e4646d8b71c52dae93a4881981b51#egg=cicflowmeter&subdirectory=cicflowmeter
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.9
Django==5.1.4
django-render-partial==0.4
djangorestframework==3.15.2
dulwich==0.21.7
fastjsonschema==2.21.1
filelock==3.16.1
idna==3.7
importlib_metadata==8.5.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
joblib==1.4.2
keyring==24.3.1
lxml==5.3.0
more-itertools==10.5.0
msgpack==1.1.0
numpy==1.26.2
packaging==23.2
pexpect==4.9.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.4.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
psutil==6.1.0
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.2.0
pyshark==0.6
pytest==8.0.0
RapidFuzz==3.10.1
requests==2.31.0
requests-toolbelt==1.0.0
scapy==2.5.0
scikit-learn==1.5.2
scipy==1.11.4
shellingham==1.5.4
sqlparse==0.5.2
termcolor==2.5.0
threadpoolctl==3.5.0
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
urllib3==2.2.1
virtualenv==20.28.0
xattr==1.1.0
zipp==3.21.0
